Added the last functionality to RecruitmentDAO and its components
Added the functionality of getting the info about an application getApplication(), and submitting a decision that is related to an application submitApplicationDecision(). In addition to some private function needed for the functionality getRoleID(), checkApplication. Added private signupRecruiter() to be able to add Recruiters to the database but it is used internally. Added a missing check in the registerNewApplication() and fixed minor line of a return.